marketplace/requirements.txt
Flask>=2.3,<3.0
Werkzeug>=2.3,<3.0
pymongo>=4.6,<5.0
python-dotenv>=1.0.0,<2.0.0
reportlab>=4.0.0,<5.0.0
email-validator>=2.0.0,<3.0.0
pymongo[srv]>=4.6,<5.0
